Initial iOS to Micropython BLE example
This example uses BLE to connect to a micropython board from iOS and allows the mobile app to set the WiFi credentials on the WifiKit32 (ESP32) board.

 The UI was written in SwiftUI.  SwiftUI is still a bit of a mystery/blackbox/early-adoption/performance when it comes to tackling things like: property change notification, UI navigation, split panel, looking good on the ipad/mac, not pre-loading views in the nativationlinks.

During my testing of the BLE on Micropython, it has been very solid and consistent.  The only issue I have seen is once you are connected to a peripheral, the iOS client needs a bit of a delay before it will accept any data.  While very small amount of time is needed, you can lose information if you are expecting it.

When setting an invalid password on the SSID, the micropython board would go into a spin (in the background, but you can see the output if you are connected via the UART) and would never give back the status of "network.STAT_WRONG_PASSWORD".

It is also worth nothing that you need to inspect the python module to find out what is really supported.  For example, bytearray does not have the 'index' method.  Also, the list of the 'network.STAT_*' are different from what is posted on the latest documentation on the micropython org site.

Overall I feel that with a little more work (and maybe the navigation bugs fixed in SwiftUI) that this could be a viable option for setting up Wifi on the WifiKit32 (ESP32) board from a mobile app.